Create learn about our ratings modal component
